chore: separete funciton in different file
