+++ backend/functions/disconnect.ts
@@ -1,0 +1,5 @@
+import { success } from '../libs/response';
+
+export const main = async () => {
+    return success();
+};
